refactor: Refactor problem files: replace int with size_t for index variables to improve type safety; update CMakeLists.txt to suppress specific warnings; remove unused input case file.

+++ CodeForces/Rounds/Round_1042_Div_3/problem_A.cpp
@@ -39,20 +39,20 @@ constexpr ll  LINF = 1e18;
 
 // Function to solve a single test case
 void solve() {
-  int n;
+  size_t n;
   cin >> n;
   vi a(n), b(n);
   ll total_excess = 0;
 
-  for (int i = 0; i < n; i++) {
+  for (size_t i = 0; i < n; i++) {
     cin >> a[i];
   }
 
-  for (int i = 0; i < n; i++) {
+  for (size_t i = 0; i < n; i++) {
     cin >> b[i];
   }
 
-  for (int i = 0; i < n; i++) {
+  for (size_t i = 0; i < n; i++) {
     if (a[i] > b[i]) {
       total_excess += a[i] - b[i];
     }
